@@ -7,22 +7,22 @@
     "name": "Espen Strømland"
   },
   "dependencies": {
-    "classnames": "2.2.6",
-    "lodash": "4.17.13",
-    "react": "16.8.6",
-    "react-dom": "16.8.6"
+    "classnames": "2.3.1",
+    "lodash": "4.17.21",
+    "react": "18.0.0",
+    "react-dom": "18.0.0"
   },
   "devDependencies": {
-    "@types/classnames": "2.2.8",
-    "@types/jest": "24.0.15",
-    "@types/lodash": "4.14.134",
-    "@types/react": "16.8.22",
-    "@types/react-dom": "16.8.4",
-    "@types/webmidi": "2.0.3",
-    "prettier": "1.18.2",
-    "react-scripts": "3.0.1",
-    "rimraf": "2.6.3",
-    "typescript": "3.5.2"
+    "@types/classnames": "2.3.1",
+    "@types/jest": "27.4.1",
+    "@types/lodash": "4.14.182",
+    "@types/react": "18.0.7",
+    "@types/react-dom": "18.0.0",
+    "@types/webmidi": "2.0.6",
+    "prettier": "2.6.2",
+    "react-scripts": "5.0.1",
+    "rimraf": "3.0.2",
+    "typescript": "4.6.3"
   },
   "scripts": {
     "start": "react-scripts start",
